requirements.txt
requests
pandas
redshift_connector
python-dotenv
